Some refactoring done.
